improve icon_folder deprecation handling
The new `icon_path` option takes precedence. If `icon_path` is not set,
we check for `icon_folders` and use that. If neither is set, use the
compile-time default.

Help output for each option is generated when the values are read. As
far as I can see, help output generation and reading the option values
cannot be decoupled. The result that if we want to read the value of
`icon_folders` we also have to generate its help (usage) output.

In other words, help output will differ depending on whether or
not you have set `icon_folders` and not `icon_path`. I don't see a way
around this but it isn't necessarily bad.

Also, `icon_folders` help output prints a deprecation warning.